tests/wpt/tests/css/css-borders/tentative/parsing/corner-shape-valid.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>CSS Borders and Box Decorations 4 Test: Parsing 'corner-shape' with valid values</title>
<link rel="author" title="Sebastian Zartner" href="mailto:sebastianzartner@gmail.com">
<link rel="help" href="https://drafts.csswg.org/css-borders-4/#corner-shaping">
<meta name="assert" content="This test checks that 'corner-shape' supports the values 'round' and 'angle'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>

<script>
    function test_values_for_prop(prop) {
        test_valid_value(prop, "round");
        test_valid_value(prop, "scoop");
        test_valid_value(prop, "notch");
        test_valid_value(prop, "bevel");
        test_valid_value(prop, "squircle");
        test_valid_value(prop, "straight");
        test_valid_value(prop, "superellipse(2)");
        test_valid_value(prop, "superellipse(.5)", "superellipse(0.5)");
        test_valid_value(prop, "superellipse(7)");
        test_valid_value(prop, "superellipse(0.3)");
        test_valid_value(prop, "superellipse(  0)", "superellipse(0)");
        test_valid_value(prop, "superellipse(2 )", "superellipse(2)");
        test_valid_value(prop, "superellipse(infinity)");
        test_valid_value(prop, "superellipse(calc(0.5 * 4))", "superellipse(calc(2))");
    }

    for (const v of ["top", "bottom"]) {
        for (const h of ["left", "right"]) {
            test_values_for_prop(`corner-${v}-${h}-shape`);
        }
    }
    for (const block of ["start", "end"]) {
        for (const inline of ["start", "end"]) {
            test_values_for_prop(`corner-${block}-${inline}-shape`);
        }
    }
</script>
